requirements.txt
astroid==2.11.2
autopep8==1.6.0
bcrypt==3.2.0
capstone==5.0.0rc2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
colored-traceback==0.3.0
cryptography==36.0.2
dill==0.3.4
gmpy2==2.1.2
idna==3.3
intervaltree==3.1.0
isort==5.10.1
lazy-object-proxy==1.7.1
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.7.0
Naked==0.1.31
packaging==21.3
paramiko==2.10.3
platformdirs==2.5.1
plumbum==1.7.2
psutil==5.9.0
pwntools==4.7.0
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
pyelftools==0.28
Pygments==2.11.2
pylint==2.13.4
PyNaCl==1.5.0
pyparsing==3.0.7
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
pywin32==303
PyYAML==6.0
requests==2.27.1
ROPGadget==6.7
rpyc==5.1.0
shellescape==3.8.1
six==1.16.0
sortedcontainers==2.4.0
toml==0.10.2
tomli==2.0.1
unicorn==2.0.0rc6
urllib3==1.26.9
wrapt==1.14.0
